python-backend/requirements.txt
fastapi
uvicorn
requests
mangum
python-jose[cryptography]
passlib[bcrypt]
pymongo
boto3
uuid
PyJWT
python-multipart
pinecone-client
openai
cohere==4.37 